+++ src/SadConsole.Shared/Renderers/LayeredConsole.cs
@@ -11,13 +11,12 @@ namespace SadConsole.Renderers
     /// <summary>
     /// Draws a text surface to the screen.
     /// </summary>
-    [DataContract]
-    public class ControlsConsole : Console
+    public class LayeredConsole : Console
     {
         /// <summary>
         /// Controls to render.
         /// </summary>
-        public List<Controls.ControlBase> Controls { get; set; } = new List<SadConsole.Controls.ControlBase>();
+        public List<CellSurfaceLayer> Layers { get; set; } = new List<CellSurfaceLayer>(0);
 
         /// <summary>
         /// Renders a surface to the screen.
@@ -26,44 +25,56 @@ namespace SadConsole.Renderers
         /// <param name="force">When <see langword="true"/> draws the surface even if <see cref="CellSurface.IsDirty"/> is <see langword="false"/>.</param>
         public override void Render(SadConsole.Console surface, bool force = false)
         {
+            foreach (var layer in Layers)
+            {
+                if (layer.IsDirty)
+                {
+                    force = true;
+                    break;
+                }
+            }
+
             RenderBegin(surface, force);
             RenderCells(surface, force);
 
-            foreach (var control in Controls)
-                RenderControlCells(control, surface.IsDirty || force);
+            foreach (var layer in Layers)
+                if (layer.IsVisible)
+                    RenderLayer(layer, surface, surface.IsDirty || force);
 
             //RenderControls(surface, force);
             RenderTint(surface, force);
             RenderEnd(surface, force);
         }
 
-        public void RenderControlCells(ControlBase control, bool draw = false)
+        /// <summary>
+        /// Renders a layer on top of a console.
+        /// </summary>
+        /// <param name="layer">The layer to render.</param>
+        /// <param name="drawingHost">The console that will be drawn on.</param>
+        /// <param name="draw">If <see langword="false"/>, skips rendering.</param>
+        public void RenderLayer(CellSurfaceLayer layer, SadConsole.Console drawingHost, bool draw = true)
         {
             if (!draw) return;
-            //if (control.Surface.Tint.A == 255) return;
 
-            if (control.Surface.DefaultBackground.A != 0)
-                Global.SpriteBatch.Draw(control.Parent.Font.FontImage, new Rectangle(control.Position.ConsoleLocationToPixel(control.Parent.Font), new Point(control.Width, control.Height) * control.Parent.Font.Size), control.Parent.Font.GlyphRects[control.Parent.Font.SolidGlyphIndex], control.Surface.DefaultBackground, 0f, Vector2.Zero, SpriteEffects.None, 0.2f);
-
-            for (var i = 0; i < control.Surface.Cells.Length; i++)
+            if (drawingHost.Tint.A != 255)
             {
-                ref var cell = ref control.Surface.Cells[i];
-
-                if (!cell.IsVisible) continue;
-
-                Rectangle renderRect = control.Parent.RenderRects[control.Position.ToIndex(control.Parent.Width)];
-                renderRect.Location += control.Surface.GetPointFromIndex(i).ConsoleLocationToPixel(control.Parent.Font);
-
-                if (cell.Background != Color.Transparent && cell.Background != control.Surface.DefaultBackground)
-                    Global.SpriteBatch.Draw(control.Parent.Font.FontImage, renderRect, control.Parent.Font.GlyphRects[control.Parent.Font.SolidGlyphIndex], cell.Background, 0f, Vector2.Zero, SpriteEffects.None, 0.3f);
-
-                if (cell.Foreground != Color.Transparent)
-                    Global.SpriteBatch.Draw(control.Parent.Font.FontImage, renderRect, control.Parent.Font.GlyphRects[cell.Glyph], cell.Foreground, 0f, Vector2.Zero, cell.Mirror, 0.4f);
-
-                foreach (var decorator in cell.Decorators)
+                for (var i = 0; i < layer.Cells.Length; i++)
                 {
-                    if (decorator.Color != Color.Transparent)
-                        Global.SpriteBatch.Draw(control.Parent.Font.FontImage, renderRect, control.Parent.Font.GlyphRects[decorator.Glyph], decorator.Color, 0f, Vector2.Zero, decorator.Mirror, 0.5f);
+                    ref var cell = ref layer.Cells[i];
+
+                    if (!cell.IsVisible) continue;
+
+                    if (cell.Background != Color.Transparent)
+                        Global.SpriteBatch.Draw(drawingHost.Font.FontImage, drawingHost.RenderRects[i], drawingHost.Font.GlyphRects[drawingHost.Font.SolidGlyphIndex], cell.Background, 0f, Vector2.Zero, SpriteEffects.None, 0.3f);
+
+                    if (cell.Foreground != Color.Transparent)
+                        Global.SpriteBatch.Draw(drawingHost.Font.FontImage, drawingHost.RenderRects[i], drawingHost.Font.GlyphRects[cell.Glyph], cell.Foreground, 0f, Vector2.Zero, cell.Mirror, 0.4f);
+
+                    foreach (var decorator in cell.Decorators)
+                    {
+                        if (decorator.Color != Color.Transparent)
+                            Global.SpriteBatch.Draw(drawingHost.Font.FontImage, drawingHost.RenderRects[i], drawingHost.Font.GlyphRects[decorator.Glyph], decorator.Color, 0f, Vector2.Zero, decorator.Mirror, 0.5f);
+                    }
                 }
             }
         }