+++ src/SadConsole.Shared/SerializedTypes/Console.cs
@@ -28,6 +28,9 @@ namespace SadConsole.SerializedTypes
         [DataMember] public int Height;
         [DataMember] public bool IsVisible;
         [DataMember] public bool IsPaused;
+        [DataMember] public bool UsePixelPositioning;
+        [DataMember] public bool UseMouse;
+        [DataMember] public bool UseKeyboard;
         [DataMember] public PointSerialized Position;
         [DataMember] public FontSerialized Font;
         [DataMember] public ColorSerialized Tint;
@@ -40,12 +43,15 @@ namespace SadConsole.SerializedTypes
             return new ConsoleSerialized()
             {
                 Font = surface.Font,
-                Cells = surface.Cells.Select(c => (CellSerialized) c).ToArray(),
+                Cells = surface.Cells.Select(c => (CellSerialized)c).ToArray(),
                 Width = surface.Width,
                 Height = surface.Height,
                 DefaultForeground = surface.DefaultForeground,
                 DefaultBackground = surface.DefaultBackground,
                 Tint = surface.Tint,
+                UsePixelPositioning = surface.UsePixelPositioning,
+                UseMouse = surface.UseMouse,
+                UseKeyboard = surface.UseKeyboard,
                 Position = surface.Position,
                 IsVisible = surface.IsVisible,
                 IsPaused = surface.IsPaused
@@ -62,7 +68,10 @@ namespace SadConsole.SerializedTypes
                 DefaultBackground = surface.DefaultBackground,
                 Position = surface.Position,
                 IsVisible = surface.IsVisible,
-                IsPaused = surface.IsPaused
+                IsPaused = surface.IsPaused,
+                UsePixelPositioning = surface.UsePixelPositioning,
+                UseKeyboard = surface.UseKeyboard,
+                UseMouse = surface.UseMouse,
             };
         }
     }